refactor: RegisterController 클래스 분리 및 커스텀 Exception 생성
